Reject sanitization that would result in an empty array
In this change, we reject an acceptedResourceRoles value that would
result in an empty array, rather than substitute it with a default
value. This will reduce confusion when a user unknowingly is creating a
service as the group role and tries to launch the service in a
statically reserved reservation.

JIRA Issues: MARATHON-8748